make longs

+++ ESPGeiger/src/Status.h
@@ -64,8 +64,8 @@ struct Status {
   long start = 0;
   bool warmup = true;
   char* geiger_model = "";
-  uint32_t total_clicks = 0;
-  uint32_t clicks_today = 0;
+  unsigned long total_clicks = 0;
+  unsigned long clicks_today = 0;
   float partial_clicks = 0;
   unsigned long last_send = 0;
   unsigned long last_blip = 0;
